services/consumer/requirements.txt
kafka-python==2.0.2
minio==7.2.15
psycopg2-binary==2.9.9
